backend/accounts/permissions.py
from rest_framework import permissions


class IsAdminUser(permissions.BasePermission):
    """
    Custom permission to only allow admin users.
    """
    def has_permission(self, request, view):
        return (
            request.user and
            request.user.is_authenticated and
            request.user.role == 'admin'
        )


class IsUserOrAdmin(permissions.BasePermission):
    """
    Custom permission to allow users to access their own data or admins to access all.
    """
    def has_object_permission(self, request, view, obj):
        # Admin can access everything
        if request.user.role == 'admin':
            return True
        # Users can only access their own data
        return obj == request.user
